rename types.py to dtypes.py because it causes probelems with calls like
if not isinstance(subclass, (type, types.ClassType)):
