+++ test/__main__.py
@@ -1,12 +1,9 @@
 from samplefile import *
-from treetest import *
-from eventtest import *
 from commands import *
 from links import *
 from mapping import *
 from proxy import *
-from types import *
-from dnd import *
+from dtypes import *
 from bugs import *
 from infer_type import *
 from validation import *
